[FIX] delete key
