+++ src/navigation/wallet/screens/DeleteKey.tsx
@@ -14,19 +14,18 @@ import {AppActions} from '../../../store/app';
 import {sleep} from '../../../utils/helper-methods';
 import {
   deleteKey,
+  setExpectedKeyLengthChange,
   updatePortfolioBalance,
 } from '../../../store/wallet/wallet.actions';
 import {findKeyByKeyId} from '../../../store/wallet/utils/wallet';
 import useAppSelector from '../../../utils/hooks/useAppSelector';
-import {
-  setExpectedKeyLengthChange,
-  setHomeCarouselConfig,
-} from '../../../store/app/app.actions';
+import {setHomeCarouselConfig} from '../../../store/app/app.actions';
 import {
   unSubscribeEmailNotifications,
   unSubscribePushNotifications,
 } from '../../../store/app/app.effects';
 import {useTranslation} from 'react-i18next';
+import {useAppDispatch} from '../../../utils/hooks';
 
 const DeleteKeyContainer = styled.SafeAreaView`
   flex: 1;
@@ -48,7 +47,7 @@ const DeleteKeyParagraph = styled(Paragraph)`
 const DeleteKey = () => {
   const {t} = useTranslation();
   const navigation = useNavigation();
-  const dispatch = useDispatch();
+  const dispatch = useAppDispatch();
   const homeCarouselConfig = useAppSelector(({APP}) => APP.homeCarouselConfig);
 
   const {notificationsAccepted, emailNotifications, brazeEid} = useAppSelector(
@@ -92,8 +91,8 @@ const DeleteKey = () => {
     const previousKeysLength = Object.keys(keys).length;
     const numNewKeys = Object.keys(keys).length - 1;
     const expectedLengthChange = previousKeysLength - numNewKeys;
-    dispatch(deleteKey({keyId}));
     dispatch(setExpectedKeyLengthChange(expectedLengthChange));
+    dispatch(deleteKey({keyId}));
     dispatch(
       setHomeCarouselConfig(
         homeCarouselConfig.filter(item => item.id !== keyId),
