+++ src/store/app/app.actions.ts
@@ -304,10 +304,3 @@ export const setUserFeedback = (feedBack: FeedbackType): AppActionType => ({
   type: AppActionTypes.USER_FEEDBACK,
   payload: feedBack,
 });
-
-export const setExpectedKeyLengthChange = (
-  lengthChange: number,
-): AppActionType => ({
-  type: AppActionTypes.EXPECTED_KEY_LENGTH_CHANGE,
-  payload: lengthChange,
-});
