+++ src/store/app/app.reducer.ts
@@ -38,7 +38,6 @@ export const appReduxPersistBlackList: Array<keyof AppState> = [
   'activeModalId',
   'failedAppInit',
   'brazeContentCardSubscription',
-  'expectedKeyLengthChange',
 ];
 
 export type ModalId = 'sheetModal' | 'ongoingProcess' | 'pin';
@@ -122,7 +121,6 @@ export interface AppState {
   failedAppInit: boolean;
   checkingBiometricForSending: boolean;
   hasViewedZenLedgerWarning: boolean;
-  expectedKeyLengthChange: number;
 }
 
 const initialState: AppState = {
@@ -196,7 +194,6 @@ const initialState: AppState = {
   failedAppInit: false,
   checkingBiometricForSending: false,
   hasViewedZenLedgerWarning: false,
-  expectedKeyLengthChange: 0,
 };
 
 export const appReducer = (
@@ -572,12 +569,6 @@ export const appReducer = (
       return {
         ...state,
         userFeedback: action.payload,
-      };
-
-    case AppActionTypes.EXPECTED_KEY_LENGTH_CHANGE:
-      return {
-        ...state,
-        expectedKeyLengthChange: action.payload,
       };
 
     default:
